requirements.txt
anyio==4.4.0
attrs==24.2.0
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
outcome==1.3.0.post0
psutil==6.0.0
PySocks==1.7.1
python-telegram-bot==21.5
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
